+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- New DS core component: accordion [#632](https://github.com/CartoDB/carto-react/pull/632)
+
 ## 2.0
 
 ### 2.0.0-beta.8 (2023-04-04)